+++ OneBlink.SDK/models/FormElement.cs
@@ -4,6 +4,11 @@ using System.Linq;
 
 namespace OneBlink.SDK.Model
 {
+    public enum HintPositionEnum
+    {
+        BELOW_LABEL,
+        TOOLTIP
+    }
     public class FormElement
     {
         public Guid id
@@ -224,6 +229,10 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
+        public HintPositionEnum hintPosition
+        {
+            get; set;
+        }
         public string placeholderValue
         {
             get; set;
@@ -420,6 +429,7 @@ namespace OneBlink.SDK.Model
             string defaultValue = null,
             string placeholderValue = null,
             string hint = null,
+            HintPositionEnum hintPosition = HintPositionEnum.BELOW_LABEL,
             int? minLength = null,
             int? maxLength = null,
             List<string> customCssClasses = null,
@@ -441,6 +451,7 @@ namespace OneBlink.SDK.Model
             textElement.minLength = minLength;
             textElement.maxLength = maxLength;
             textElement.hint = hint;
+            textElement.hintPosition = hintPosition;
             textElement.customCssClasses = customCssClasses;
             textElement.meta = meta;
             return textElement;
@@ -458,6 +469,7 @@ namespace OneBlink.SDK.Model
             string defaultValue = null,
             string placeholderValue = null,
             string hint = null,
+            HintPositionEnum hintPosition = HintPositionEnum.BELOW_LABEL,
             List<string> stateTerritoryFilter = null,
             List<string> customCssClasses = null,
             string meta = null
@@ -476,6 +488,7 @@ namespace OneBlink.SDK.Model
             geoscapeAddressElement.defaultValue = defaultValue;
             geoscapeAddressElement.placeholderValue = placeholderValue;
             geoscapeAddressElement.hint = hint;
+            geoscapeAddressElement.hintPosition = hintPosition;
             geoscapeAddressElement.stateTerritoryFilter = stateTerritoryFilter;
             geoscapeAddressElement.customCssClasses = customCssClasses;
             geoscapeAddressElement.meta = meta;
@@ -521,6 +534,7 @@ namespace OneBlink.SDK.Model
             bool readOnly = false,
             string defaultValue = null,
             string hint = null,
+            HintPositionEnum hintPosition = HintPositionEnum.BELOW_LABEL,
             bool isDataLookup = false,
             long? dataLookupId = null,
             bool isElementLookup = false,
@@ -542,6 +556,7 @@ namespace OneBlink.SDK.Model
             complianceElement.readOnly = readOnly;
             complianceElement.defaultValue = defaultValue;
             complianceElement.hint = hint;
+            complianceElement.hintPosition = hintPosition;
             complianceElement.isDataLookup = isDataLookup;
             if (dataLookupId.HasValue)
             {
@@ -569,6 +584,7 @@ namespace OneBlink.SDK.Model
             string defaultValue = null,
             string placeholderValue = null,
             string hint = null,
+            HintPositionEnum hintPosition = HintPositionEnum.BELOW_LABEL,
             List<string> stateTerritoryFilter = null,
             List<string> addressTypeFilter = null,
             List<string> customCssClasses = null,
@@ -588,6 +604,7 @@ namespace OneBlink.SDK.Model
             pointAddressElement.defaultValue = defaultValue;
             pointAddressElement.placeholderValue = placeholderValue;
             pointAddressElement.hint = hint;
+            pointAddressElement.hintPosition = hintPosition;
             pointAddressElement.stateTerritoryFilter = stateTerritoryFilter;
             pointAddressElement.addressTypeFilter = addressTypeFilter;
             pointAddressElement.customCssClasses = customCssClasses;
@@ -606,6 +623,7 @@ namespace OneBlink.SDK.Model
             bool readOnly = false,
             bool defaultValue = false,
             string hint = null,
+            HintPositionEnum hintPosition = HintPositionEnum.BELOW_LABEL,
             bool isDataLookup = false,
             long? dataLookupId = null,
             bool isElementLookup = false,
@@ -626,6 +644,7 @@ namespace OneBlink.SDK.Model
             booleanElement.readOnly = readOnly;
             booleanElement.defaultValue = defaultValue;
             booleanElement.hint = hint;
+            booleanElement.hintPosition = hintPosition;
             booleanElement.isDataLookup = isDataLookup;
             if (dataLookupId.HasValue)
             {
@@ -652,6 +671,7 @@ namespace OneBlink.SDK.Model
             bool readOnly = false,
             dynamic defaultValue = null,
             string hint = null,
+            HintPositionEnum hintPosition = HintPositionEnum.BELOW_LABEL,
             Boolean useGeoscapeAddressing = false,
             string titleLabel = null,
             string familyNameLabel = null,
@@ -693,6 +713,7 @@ namespace OneBlink.SDK.Model
             civicaNameRecordElement.readOnly = readOnly;
             civicaNameRecordElement.defaultValue = defaultValue;
             civicaNameRecordElement.hint = hint;
+            civicaNameRecordElement.hintPosition = hintPosition;
             civicaNameRecordElement.useGeoscapeAddressing = useGeoscapeAddressing;
             civicaNameRecordElement.titleLabel = titleLabel;
             civicaNameRecordElement.familyNameLabel = familyNameLabel;
@@ -801,6 +822,7 @@ namespace OneBlink.SDK.Model
             Guid? id = null,
             bool readOnly = false,
             string hint = null,
+            HintPositionEnum hintPosition = HintPositionEnum.BELOW_LABEL,
             string storageType = null,
             bool restrictFileTypes = false,
             List<string> restrictedFileTypes = null,
@@ -825,6 +847,7 @@ namespace OneBlink.SDK.Model
             filesElement.id = initialiseId(id);
             filesElement.readOnly = readOnly;
             filesElement.hint = hint;
+            filesElement.hintPosition = hintPosition;
             filesElement.storageType = storageType;
             filesElement.restrictFileTypes = restrictFileTypes;
             filesElement.restrictedFileTypes = restrictedFileTypes;