+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+-   `hintPosition` to `FormElement`
+
 ## 4.1.1 (2023-03-13)
 
 ### Added